[Lab2] Ex. 5: implemented page functions
Line 'page_check() succeeded!' is displayed according to test with this
patch. Note in particular that page_insert() is not implemented in terms
of page_remove(), as instructed in the comments. It turns out that we
have to call pgdir_walk() in page_remove() to do something cannot easily
be done with page_remove(): we need to determine whether the page to
be inserted is exactly the same one already there and proceed accordingly.
The idea of decrementing ref of the old page while incrementing that of
the new one won't work since we won't know what to do with refs when
pgdir_walk() fails.

This exercise, like the previous one, is very hard to get correct in
the first place.

Checklists that may help debug page problems like in this exercise:
1. Whether page's ref is incremented when it is used (e.g., inserted)?
2. Whether C macros, e.g., those defined in <inc/queue.h> and transformers
   among kernel (virtual) addresses, physical addresses, page addresses,
   etc., are used with proper arguments?