[lab3] Amend env icode loading and running
The previous code is buggy; we made that up because of lack of complete
understanding of the whole thing. The phenomenon of the bug is wierd,
hard-to-spot, undesired change to page table of specific environment(s).
We are yet to tell a full story about the bug.

This commit is made by 'stealing' code from work by zhangchitc@gmail.com.