[lab4] Ex 1. Impl. round-robin scheduling
The commit is self-evident; kern/init.c is modified to demonstrate that the
scheduling code works.

While doing this commit, a strange bug prevails that cause page fault in
switch-case dispatch code in vprintfmt(). This bug turned out to be related
to the way environment pages are set up. It was previously introduced in
lab3, and hence a fix was made there. Our lab4 code is then rebased against
that commit: that's why the fix is newer in date than the first commit in
lab4.