Fixed page_insert() for inserting the old page
Previously, we made a small 'optimization' that does nothing when the page
to inserted is the same one already existing there. This turned out to be
disastrous: lib/fork.c intends to insert the same old page again just to
change page permisssion.

Now the code is back to the standard way of manipulating reference counts.

This bug is also known as another 10000RMB one, which cost us 25 days to
debug and fix.