[lab5][ex1] Grant the fs process I/O priviledge
Since the fs process is always the process 1 (immediately after user/idle),
what that is needed to do is to grant the priviledge at env_create().

By the way, the processes to run in kern/init.c is replaced, now that the
user programs are tested OK with the newly imported fs code and that we are
working on new code in lab5.