[lab5][ex4] Impl. the rest 'top-level' functions
One special modification is that in alloc_block(), the memory for the newly
allocated block is zeroed. This block, being just allocated, must contain
garbage in both the in-memory version and the in-disk version. Without the
zeroing, the content of this block in memory now mysteriously turns garbage
and disagrees with the in-disk version.