[lab6] [ex3] Fixed fd/pipe page ref race
This 'fix' is part of the exercise; it addresses the first category of
race conditions: file descriptor/pipe page reference updating race.
Mainly, the fd is unmapped in advance in pipe_close() (in close() in
lib/fd.c, fd is unmapped after fddata/pipe is unmapped) and mapped last
in dup().

The programs user/primespipe and user/testpiperace are tested OK. Be
warned that user/primespipe may take plenty of time to run.