[lab6][ex-5] Re-work pipe r/w based on naive mutex
Introducing mutex as kernel facility and exposing it via system calls has
been seen as one way of solving the read/write race conditions thoroughly.
The implementation here is naive and ineffecient, but works. Then piperead()
and pipewrite() are changed to use mutex (they are most difficult to get
right !!!); the shell pipe is now more reliable.

A few bugs are fixed while trying to make the shell commands listed in the
lab work:
1) The 'offset' parameters in piperead() and pipewrite() are now absolute,
not relative to read position or write position, as we previously thought.
The problem is only observed when we came up with the incorrect versions and
examined the output; after that we have to begin again.
2) Fixed sys_page_alloc() to properly zero the memory being allocated. This
is discovered because of wierd pipe contents when running some program for a
second time.
3) Fixed lib/spawn.c:init_stack() to properly initialize argv. The problem
was observed when we found that if we run 'echo 1 2 3', the program outputs
'3 3 3'.
4) While redirecting output to file in sh via '>', truncate the file size
to zero.

Kernel development is hard; debugging with printf is even harder!!! That's
why we were forced to write a lot of cprintf clauses. We've decided not to
delete them, unlike what we previously did, in the hope that they might
become useful again (e.g., when more bugs are spotted).