[lab6][ex-5] challenge: added support for '&'
A command with a terminal '&' is meant to run in background. That is,
the shell will return immediately and wait for the next command, instead
of waiting for the command to exit.

Note that currently the support is incomplete. For example, when we run
'ls &', we are immediately back to the command prompt as expected, but
the ls command does not run until we run several other commands like
echo. Apparently, the process scheduling algorithm is to blame.