Fix kernenl loop that fails background command
Motivation:
The background command does not run until another sync process is run,
giving the impression that there is something wrong with the process
scheduling algorithm.

But better looked into, actually the userland processes are starving,
as the kernel is busy polling console input as part of processing
sys_cgetc().

Changes:
The sys_cget() is changed to return immediately even there is no console
input yet. The caller in cons_read() lib/console.c correctly embeds the
syscall along with a sys_yield() within an infinite loop, so no change
is needed for it.

Additional trivial changes are made to the user/sh.c code.

If additional performance is needed, however, the infinite-polling
mechanism shall be replaced with an async version.

Results:
Now background command works as expected. The rule that kernel shall
never fall into infinite loop is enforced globally.